requirements.txt
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==4.2.21
google-generativeai==0.8.3
openai==1.54.4
isort==6.0.1
mccabe==0.7.0
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
python-decouple==3.8